List#to_s now only generates the number of list items to be displayed.
It's a shame that this logic has to live here (previously Rembrandt) was
able to take a list of any length and display it based on what was
selected). However, the new way of displaying lists (which is the way
the current iPods do it) requires that either Rembrandt or the List
class keeps some sort of display state around. The List seems like the
best place.