Added a card cache. This caches Card instances based on the card stack.
This allows for a single Card class to have multiple instances cached,
based on the Cards that were accessed prior to it. This is to allow
things like an instance of ArtistsCard to cache the full list of
artists, but another instance of ArtistsCard to cache only a list of
artists for a specific genre.