The params hash is now set in Spandex::Card rather than subclasses.
This results in more consistency and less code for the children.
Card#after_load is called once the params are set. This is useful for
the children to maipulate the params hash if they need to.

Added before_show and after_show methods. Those and show are called
using call_show_chain.